subscription-registry: Optimized by reducing number of setTimeout calls when `broadcastTimeout > 0`.
Since `broadcastTimeout` is a soft limit we can simply use one global timeout instead of allocating one timeout for each topic. With e.g. a 10ms timeout and 10 000 topics we would go from a worst case of 1 000 000 allocs/s to 100 allocs/s.
With this change `broadcastTimeout` goes from a hard limit to a soft limit where the average buffer time will be `broadcastTimeout/2` and the longest `broadcastTimeout`.