feat: holding a reference to authenticated sockets
This commit is a bit hard to explain, since it doesn't actually provide any value
as of yet within the opensource branch. However, we are working on multiple plugins
which architecture will go into opensouce in the next couple of weeks that will allow
multiple different connection endpoints to be added.