fix: logger method overriding
to allow use of the logger.warn and logger.error methods in older loggers, we provide defaults
during config initialization. however, the code was binding to a null object leading to
incorrect behaviour. this commit also ensures that these methods may be overridden by a logger
that wishes to implement them.