add optional config and config loading via string.
There was a problem in the way deepstream used the configuration object. If I want to set some of the parameter, I had also set all the other, because there weren't optional. So I checked if the algorithm worked like 3.1.1. and it themes. That mean, optional parameter won't break things. There was also a problem to set string as the config, like for files. I also fixed this.