Fix of configuration issues
There are some isssues, I found while setting config options to optional. I added some checks, but it would be better to refactor how your configuration object works. It is not necessary, that everyone get your config object. Isolation against configuration properties and the modules, which will import them, could help to reduce complexity and side effects. There is a chance, that they can have an effect on the behavior. I hope this will be a topic for your company in the next time.