refactor: Improving performance by converting to ES6 Objects
Dynamic object lookups are slower than map.get, specially when you look at operations such
as has and delete. Its generally also cleaner due to typings.

We also snuck the removal of the ping messages in this commit. The goal is for the client
to send any message within a certain time, and if they don't because they are idling send
a ping instead. This is probably a small step back as the client doesn't keep track of the
server status this way, however we can add that in post V4 release.