refactor: listening overhaul
The last two commits got rid of most of the logic related to listening. The reason this
was done was to improve debugging massively. Alot of log lines are still needed but its
a huge step in the right direction.

The previous active listening used a very TDD approach. This means that each scenario was
added as an extra case and while it allowed us to deal with insane edge cases like:

provider 1 rejects, 2 timeouts out and is stored to see if it eventually accepts, 3 rejects, 2 accepts, 4 rejects, two is chosen

we now have

provider 1 rejects, 2 times out and is told it was disregarded, 3 rejects. X seconds later 1 rejects, 2 accepts, two is chosen

while this seems like a tiny change it simplifies logic massively.

Clustering also changed from this:

- Get a subscription on server 1
- server 1 retrieves lock
- server 1 tries to get local provider
- no local provider, server 1 figures which servers in cluster to 'lead'
- server 1 asks each server in cluster that has potential providers to try (sequentially)
- if a server accepts, stops leading and releases lock
- if a server rejects, tries next one till all run out

To this:

- Get a subscription server 1
- all servers retrieve a lock (usually its a race between the cluster leader and node the subscription was made on)
- server with lock tries to get local provider
- no local provider, server stores subscription name with time and tells the cluster it failed
- cluster then repeats previous steps
- eventually a node finds a provider, or the all ignore the message due to the time based debounce
- every X amount of seconds each server gives trying to match a provider another shot

This is a bit more undeterministic (read, OMG SHOTGUN!) and can lead to some issues where the server subscription isn't
always on the same machine as the provider, however, that isn't extremely likely. Also keep in mind not accepting a request
in time is generally broken code and shouldn't be guarded against so closely!