Feature: record notify action
Ability to inform deepstream that updates (to both cache and storage!)
occured. This is super useful if you want to maintain the db via a non
deepstream component. Still a little bit of work left to nuke everything
that changed remotely from the cache first so that it truly works with
db syncing